changes in argparse, using linsi

+++ Pipelign.py
@@ -15,7 +15,7 @@ from Bio import SeqIO
 import tempfile
 import time
 
-
+#************************************
 class Struct:
   '''
    - this class creates a structure of the parameter values needed for different functions.
@@ -25,10 +25,122 @@ class Struct:
     for k, v in kwargs.items():
       setattr(self,k,v)
       
+#*************************************      
+
+#*************************************
 class MyStruct(Struct):
   pass
 
-#********************************************************************
+#*****************************************
+
+#******************************************
+def lengthThreshold(x):
+  x = float(x)
+  if x < 0.0 or x > 1.0:
+    raise argparse.ArgumentTypeError('%r not in range [0.0, 1.0]' % (x,))
+  return x
+#***************************************
+
+#****************************************
+def genCodeLimit(x):
+  x = int(x)
+  gCodes = list(1,2,3,4,5,6,9,10,11,12,13,14,16,21,22,23,24,25)
+  if x not in gCodes:
+    raise argparse.ArgumentTypeError('%r is not a valid genetic code' % (x,))
+  else:
+    return x
+
+#****************************************
+
+def cZip(cDir,tName):
+  '''
+  creates a zip file of the temporary directory
+  '''  
+  os.chdir(cDir)
+  
+  zName = 'pipelign.' + time.strftime('%Y-%m-%d-%H%M%S') 
+  shutil.make_archive(zName,'zip',tName)
+  print('\nArchive for all temporary files created in %s.zip\n' % zName)
+  sys.exit()
+
+#******************************************
+
+#***********************************************************************
+
+def ginsi(seqFile,alnFile,thread,mIterLong,log):
+  '''
+    - aligns sequences using MAFFT's G-INS-i method
+  '''
+  
+  lh = open(log,'a')
+  
+  cl = 'mafft --globalpair --thread %d --maxiterate %d --preservecase %s > %s' % (thread, mIterLong,seqFile, alnFile)
+
+  try:
+    subprocess.check_call(cl, shell=True, stdout=lh, stderr=lh)
+  except subprocess.CalledProcessError as e:
+    sys.exit(e)
+
+  lh.close()
+#***********************************************************************
+
+#***********************************************************************
+
+def linsi(seqFile,alnFile,thread,mIterLong,log):
+  '''
+    - aligns sequences using MAFFT's L-INS-i method
+  '''
+  
+  lh = open(log,'a')
+  
+  cl = 'mafft --localpair --thread %d --maxiterate %d --preservecase %s > %s' % (thread, mIterLong,seqFile, alnFile)
+
+  try:
+    subprocess.check_call(cl, shell=True, stdout=lh, stderr=lh)
+  except subprocess.CalledProcessError as e:
+    sys.exit(e)
+
+  lh.close()
+#***********************************************************************
+
+#***********************************************************************
+
+def fftnsi(seqFile,alnFile,thread,mIterLong,log):
+  '''
+    - aligns sequences using MAFFT's FFT-NS-i method
+  '''
+  
+  lh = open(log,'a')
+  
+  cl = 'mafft --thread %d --maxiterate %d --preservecase %s > %s' % (thread, mIterLong,seqFile, alnFile)
+
+  try:
+    subprocess.check_call(cl, shell=True, stdout=lh, stderr=lh)
+  except subprocess.CalledProcessError as e:
+    sys.exit(e)
+
+  lh.close()
+#***********************************************************************
+
+#***********************************************************************
+
+def fftns2(seqFile,alnFile,thread,mIterLong,log):
+  '''
+    - aligns sequences using MAFFT's FFT-NS-i method
+  '''
+  
+  lh = open(log,'a')
+  
+  cl = 'mafft --thread %d --retree 2 --preservecase %s > %s' % (thread, seqFile, alnFile)
+
+  try:
+    subprocess.check_call(cl, shell=True, stdout=lh, stderr=lh)
+  except subprocess.CalledProcessError as e:
+    sys.exit(e)
+
+  lh.close()
+#***********************************************************************
+
 
 #************************************************************************
 
@@ -53,17 +165,6 @@ def deAlign(iFile, dFile):
   print('\tGapless sequence file written in <%s>' % dFile)
   
 #*************************************************************************
-
-def cZip(cDir,tName):
-  '''
-  creates a zip file of the temporary directory
-  '''  
-  os.chdir(cDir)
-  
-  zName = 'pipelign.' + time.strftime('%Y-%m-%d-%H%M%S') 
-  shutil.make_archive(zName,'zip',tName)
-  print('\nArchive for all temporary files created in %s.zip\n' % zName)
-  sys.exit()
 
 #************************************************************************
 
@@ -257,25 +358,6 @@ def addClusterNumberToReps(repName,lstFile,outFile):
 
 #***********************************************************************
 
-def ginsi(seqFile,alnFile,thread,mIterLong,log):
-  '''
-    - aligns sequences using MAFFT's G-INS-i method
-  '''
-  
-  lh = open(log,'a')
-  
-  cl = 'mafft --globalpair --thread %d --maxiterate %d --preservecase %s > %s' % (thread, mIterLong,seqFile, alnFile)
-      #cl = 'clustalo -i %s -o %s' % (cName, aName)
-  try:
-    subprocess.check_call(cl, shell=True, stdout=lh, stderr=lh)
-  except subprocess.CalledProcessError as e:
-    sys.exit(e)
-
-  lh.close()
-#***********************************************************************
-
-#***********************************************************************
-
 def addFragments(fName, aName, oName, thread, log):
   '''
     - add fragments to the alignment using MAFFT's --add
@@ -304,7 +386,8 @@ def makeClusterRepsAlignment(repFile,outFile,thread,mIterL):
 
   print('\nAligning cluster representative sequences')
   
-  ginsi(repFile,outFile,thread,mIterL,'clsRepAln.log')
+  #ginsi(repFile,outFile,thread,mIterL,'clsRepAln.log')
+  fftnsi(repFile,outFile,thread,mIterL,'clsRepAln.log')
   
   print('\tAlignment written in %s' % outFile)
 #***********************************************************************
@@ -354,15 +437,8 @@ def alnFullSequenceClusters(nClusters, thread,mIterL):
     
     if len(seqs) > 1:
       print('\tAligning cluster %d of %d sequences' % (i+1,len(seqs)))
-      ginsi(cName,aName,thread,mIterL,log)
-      '''
-      cl = 'mafft --globalpair --thread %d --maxiterate 1000 --preservecase %s > %s' % (thread, cName, aName)
-      #cl = 'clustalo -i %s -o %s' % (cName, aName)
-      try:
-        subprocess.check_call(cl, shell=True, stdout=lh, stderr=lh)
-      except subprocess.CalledProcessError as e:
-        sys.exit(e)
-      '''  
+      #ginsi(cName,aName,thread,mIterL,log)
+      linsi(cName,aName,thread,mIterL,log)
     else:
       shutil.copyfile(cName,aName)
 
@@ -370,7 +446,7 @@ def alnFullSequenceClusters(nClusters, thread,mIterL):
 #***********************************************************************  
 
 #***********************************************************************  
-def makeHMMdb(nClusters,cDir,tName,thread,lFile):
+def makeHMMdb(nClusters,cDir,tName,thread,lFile,alpha):
   '''
     Create profile HMMs for each of the full length alignments 
     Create HMM_DB from the pHMMs
@@ -408,8 +484,11 @@ def makeHMMdb(nClusters,cDir,tName,thread,lFile):
         aName = 'temp.aln'
     
     #**************
+    if alpha == 'nt':
+      cl = 'hmmbuild --dna --cpu %d %s %s' % (thread, hName, aName)
     
-    cl = 'hmmbuild --cpu %d %s %s' % (thread, hName, aName)
+    elif alpha == 'aa':
+      cl = 'hmmbuild --amino --cpu %d %s %s' % (thread, hName, aName)
     try:
       subprocess.check_call(cl,shell=True,stdout=lh,stderr=lh)
     except subprocess.CalledProcessError as e:
@@ -541,15 +620,17 @@ def parseHMMsearchResult(nClusters,fragFile,res,keepFrag):
 
     if keepFrag:
       print('But will be added to the final alignment')
-      print('\nUse the flag "-f 0" to exclude these from the final alignment')
+      #print('\nUse the flag "-f 0" to exclude these from the final alignment')
+      return True
     
     else: # user did not explicitely asked to add orphan fragments  
+      print('Please use "-f" flag to include fragments without clusters')
       while True:
         choice = input('Do you want to add these sequences to the alignment (y/n)? ')
         if choice.strip() == 'y' or choice.strip() == 'Y':
-          return 1
+          return True
         elif choice.strip() == 'n' or choice.strip() == 'N':
-          return 0
+          return False
         else:
           print('\nIncorrect choice. Please enter "y(Y)" or "n(N)"')
           continue
@@ -575,14 +656,6 @@ def addFragmentsToClusters(nClusters, thread):
       if os.path.exists(fName) and os.stat(fName).st_size > 0:
         print('\nAdding fragments from <%s> to <%s>' % (fName,aName)) 
         addFragments(fName, aName, oName, thread, log)
-        '''
-        cl = 'mafft --preservecase --thread %d --addfragments %s %s > %s' % (thread, fName, aName, oName)
-        
-        try:
-          subprocess.check_call(cl,shell=True,stdout=lh,stderr=lh)
-        except subprocess.CalledProcessError as e:
-          sys.exit(e)
-        '''
       else:
         shutil.copy(aName,oName)
 
@@ -597,7 +670,7 @@ def mergeClusters(nClusters,outFile,addNClusters,thread,mIterM):
   
   print('\nMerging all cluster alignments together')
   
-  if nClusters < 2 and addNClusters == 0: 
+  if nClusters == 1 and not addNClusters: 
     shutil.copy('cls.0.aln',outFile)
     return
   
@@ -605,6 +678,7 @@ def mergeClusters(nClusters,outFile,addNClusters,thread,mIterM):
   catText = 'cat '
   mTab = ''
   
+  '''
   for i in range(nClusters):
     cName = 'cls.' + str(i) + '.aln'
     catText += cName + ' '
@@ -616,25 +690,60 @@ def mergeClusters(nClusters,outFile,addNClusters,thread,mIterM):
     mTab += '\n'   
   #print(mTab)
   
-  if addNClusters == 1 and os.path.exists('frag.noClusters.fas'):
+  if addNClusters and os.path.exists('frag.noClusters.fas'):
     catText += 'frag.noClusters.fas ' 
   
   catText += '> merge'
   #print(catText)
+  '''
+  #*********
+  oSeqs = [] # contains orphan long sequences and fragments
+  mFlag = False
   
-  lh = open('merge.log','w')
-  try:
-    subprocess.check_call(catText,shell=True,stdout=lh,stderr=lh)
-  except subprocess.CalledProcessError as e:
-    sys.exit(e)
+  for i in range(nClusters):
+    cName = 'cls.' + str(i) + '.aln' # name of the cluster alignment file
+    seqs = list(SeqIO.parse(cName,'fasta')) # reading the alignment
+    if len(seqs) > 1: # if more than one sequence in the alignment
+      catText += cName + ' ' # add file name to the cat string 
+      for k in range(len(seqs)): # write numbers of sequences in line
+        mTab += str(seqCount) + ' '
+        seqCount += 1
+      mTab += '\n'
+      mFlag = True
+    else:
+      oSeqs.append(seqs[0]) # if only one sequence in the alignment, write to orphan file
 
-  fh = open('subMSAtable','w')
-  fh.write(mTab)
-  fh.close()
+  if addNClusters and os.path.exists('frag.noClusters.fas'): # if there exists orphan fragments
+    feqs = list(SeqIO.parse('frag.noClusters.fas','fasta'))
+    for seq in feqs:
+      oSeqs.append(seq) # add fragments to orphan seqs file
+      #print('test')
   
+  if len(oSeqs) > 0: # if at least one orphan sequence present
+    SeqIO.write(oSeqs,'orphanSeqs.fasta','fasta') # write orphan long and fragments in file
+    catText += 'orphanSeqs.fasta ' 
   
-  cl = 'mafft --preservecase --thread %d --localpair --maxiterate %d --merge subMSAtable merge > out.aln' % (thread,mIterM)
+  catText += '> merge'
+
+  lh = open('merge.log','w')
+    
+  #*********
+  if mFlag: # at least one alignment contains more than one sequences
+    try:
+      subprocess.check_call(catText,shell=True,stdout=lh,stderr=lh)
+    except subprocess.CalledProcessError as e:
+      sys.exit(e)
+
+    fh = open('subMSAtable','w')
+    fh.write(mTab)
+    fh.close()
   
+    cl = 'mafft --preservecase --thread %d --localpair --maxiterate %d --merge subMSAtable merge > out.aln' % (thread,mIterM) # uses L-INS-i
+    #cl = 'mafft --preservecase --thread %d --maxiterate %d --merge subMSAtable merge > out.aln' % (thread,mIterM) # uses FFT-NS-i
+  
+  else:
+    cl = 'mafft --preservecase --thread %d --localpair --maxiterate %d orphanSeqs.fasta > out.aln' % (thread,mIterM) # if no alignment
+    
   try:
     subprocess.check_call(cl,shell=True,stdout=lh,stderr=lh)
   except subprocess.CalledProcessError as e:
@@ -645,6 +754,7 @@ def mergeClusters(nClusters,outFile,addNClusters,thread,mIterM):
     #print(outFile)
   
   lh.close()
+  
 #************************************************************************
     
 #************************************************************************  
@@ -656,15 +766,16 @@ def getArguments():
   parser = argparse.ArgumentParser(description="Pipelign: creates multiple sequence alignment from FASTA formatted sequence file", formatter_class=argparse.RawDescriptionHelpFormatter)  
   parser.add_argument('-i', '--input', required=True, help="Input sequence file")
   parser.add_argument('-o', '--output', required=True, help="Output alignment file")
-  parser.add_argument('-t', '--thr', nargs='?', const=0.5, type=float, help="Length threshold for full sequences", default=0.5)
-  parser.add_argument('-c', '--code', nargs='?', const=1, type=int, help="Genetic code for translation",default=1)
+  parser.add_argument('-t', '--thr', type=lengthThreshold, help="Length threshold for full sequences", default=0.5)
+  parser.add_argument('-c', '--code', type=int, help="Genetic code for translation",default=1, choices=[1,2,3,4,5,6,9,10,11,12,13,14,16,21,22,23,24,25])
   parser.add_argument('-a', '--alphabet', required=True, help='Input sequences can be DNA/Protein', choices=['nt','aa'], default='nt')
-  parser.add_argument('-f', '--keepFrag', nargs = '?', const = 1, type = int, help='Add fragments without clusters', choices=[0,1],default=1)
-  parser.add_argument('-z', '--mZip', nargs = '?', const = 0, type = int, help='Create zipped temporary files', default=0)
+  parser.add_argument('-f', '--keepFrag', help='Add fragments without clusters', action="store_true")
+  parser.add_argument('-z', '--mZip', help='Create zipped temporary files', action="store_true")
   parser.add_argument('-p', '--simPer', nargs='?', const=0.8, type=float, help="percent sequence similarity for clustering", default=0.8)
   parser.add_argument('-q', '--thread', nargs='?', const=1, type=int, help="Number of CPU to use for multithreads", default=1)
   parser.add_argument('-s', '--mIterateLong', nargs='?', const=1000, type=int, help="Number of iterations to refine long alignments", default=1000)
   parser.add_argument('-m', '--mIterateMerge', nargs='?', const=100, type=int, help="Number of iterations to refine merged alignment", default=100)
+  parser.add_argument('-d', '--tempDirPath', required=False, help="Path for temporary directory",default=None)
   
   
   args = parser.parse_args()
@@ -694,7 +805,8 @@ if __name__=="__main__":
       mIterM = args.mIterateMerge,
       fragEmpty = 1,
       longName = 'long.fas',
-      fragName = 'frag.fas')
+      fragName = 'frag.fas',
+      tempDirPath = args.tempDirPath)
   
   
   cDir = os.getcwd() # save current working directory
@@ -703,19 +815,35 @@ if __name__=="__main__":
   
   #'''
   # create temporary directory
-  try:
-    tempDir = tempfile.TemporaryDirectory() # create temporary directory to hold intermediary files
-    tFileName = tempDir.name + '/' + tName1 
-    shutil.copyfile(mArgs.inFile,tFileName) # copy input file to temporary directory 
   
-  except OSError:
-    msg = '\n\nThe sequence file "%s"  cannot be found.' % mArgs.inFile
-    msg += '\n\t**Please run Pipelign again with correct input file name**'
-    msg += '\n\tPipelign is exiting.\n'
-    sys.exit(msg)
+  if mArgs.tempDirPath is None:
+    try:
+      tempDir = tempfile.TemporaryDirectory() # create temporary directory to hold intermediary files
+      tFileName = tempDir.name + '/' + tName1 
+      shutil.copyfile(mArgs.inFile,tFileName) # copy input file to temporary directory 
+      tName = tempDir.name
+  
+    except OSError:
+      msg = '\n\nThe sequence file "%s"  cannot be found' % mArgs.inFile
+      msg += '\n\t**Please run Pipelign again with correct input file name**'
+      msg += '\n\tPipelign is exiting\n'
+      sys.exit(msg)
+  
+  else:
+    if os.path.exists(mArgs.tempDirPath):
+      tempDir = mArgs.tempDirPath + '/' + 'pipelign.' + time.strftime('%Y-%m-%d-%H%M%S')
+      try:
+        os.mkdir(tempDir)
+      except OSError:
+        sys.exit('\nCould not create temporary directory. Please try again')
+      tFileName = tempDir + '/' + tName1
+      shutil.copyfile(mArgs.inFile,tFileName) # copy input file to temporary directory 
+      tName = tempDir
+    else:
+      sys.exit('\nPath for temporary directory does not exists. Please run again with correct path.')
   
   print('\nPipelign will align sequences in: <%s>' % mArgs.inFile)
-  tName = tempDir.name
+  
   
   #change current working directory to the temp
   os.chdir(tName)
@@ -730,16 +858,19 @@ if __name__=="__main__":
   
   makeClusterRepsAlignment('clsReps.fas','clsReps.aln',mArgs.thread,mArgs.mIterL)
   
-  makeIQTree('clsReps.aln',mArgs.thread)
+  if numClusters > 2:
+    makeIQTree('clsReps.aln',mArgs.thread)
   
-
+  else:
+    print('\nNumber of cluster representative(s) is %d. Phylogenetic tree can not be built' % numClusters)
+    
   #print('\n\nNumber of clusters %d' % numClusters)
   alnFullSequenceClusters(numClusters, mArgs.thread,mArgs.mIterL)
   
   if not mArgs.fragEmpty:
     lFile = 'hmmer.log'
     oFile = 'hmm.out'
-    makeHMMdb(numClusters,cDir,tName,mArgs.thread,lFile)
+    makeHMMdb(numClusters,cDir,tName,mArgs.thread,lFile,mArgs.alphabet)
     searchHMMdb(lFile,mArgs.thread,mArgs.fragEmpty,mArgs.alphabet,oFile)
     
     #numClusters = 2
@@ -755,8 +886,12 @@ if __name__=="__main__":
   
   # create zipped file for temporary directory if -z 1 is used
   if args.mZip:
-    zName = 'pipelign.' + time.strftime('%Y-%m-%d-%H%M%S') 
-    shutil.make_archive(zName,'zip',tempDir.name)
+    if mArgs.tempDirPath is None:
+      zName = 'pipelign.' + time.strftime('%Y-%m-%d-%H%M%S') 
+    else:
+      zName = tName
+    
+    shutil.make_archive(zName,'zip',tName)
     print('\nArchive for all temporary files created in %s.zip\n' % zName)
   
 #************************************************************
